Changes for running validator in docker (#1469)
* Command to dump validator addresses from network config

* Option to create genesis config without doing actual genesis

* Add --listen-address option to validator

+++ sui/src/sui_commands.rs
@@ -42,10 +42,22 @@ pub enum SuiCommand {
         #[clap(long)]
         config: Option<PathBuf>,
     },
+    #[clap(name = "network")]
+    Network {
+        #[clap(long)]
+        config: Option<PathBuf>,
+        #[clap(short, long, help = "Dump the public keys of all authorities")]
+        dump_addresses: bool,
+    },
     #[clap(name = "genesis")]
     Genesis {
         #[clap(long, help = "Start genesis with a given config file")]
         from_config: Option<PathBuf>,
+        #[clap(
+            long,
+            help = "Build a genesis config, write it to the specified path, and exit"
+        )]
+        write_config: Option<PathBuf>,
         #[clap(long)]
         working_dir: Option<PathBuf>,
         #[clap(short, long, help = "Forces overwriting existing configuration")]
@@ -80,10 +92,33 @@ impl SuiCommand {
                     .wait_for_completion()
                     .await
             }
+            SuiCommand::Network {
+                config,
+                dump_addresses,
+            } => {
+                let config_path = config
+                    .clone()
+                    .unwrap_or(sui_config_dir()?.join(SUI_NETWORK_CONFIG));
+                let config: NetworkConfig = PersistedConfig::read(&config_path).map_err(|err| {
+                    err.context(format!(
+                        "Cannot open Sui network config file at {:?}",
+                        config_path
+                    ))
+                })?;
+
+                if *dump_addresses {
+                    for auth in config.authorities.iter() {
+                        let addr = SuiAddress::from(auth.key_pair.public_key_bytes());
+                        println!("{}:{} - {}", auth.host, auth.port, addr);
+                    }
+                }
+                Ok(())
+            }
             SuiCommand::Genesis {
                 working_dir,
                 force,
                 from_config,
+                write_config,
             } => {
                 let sui_config_dir = &match working_dir {
                     // if a directory is specified, it must exist (it
@@ -102,14 +137,15 @@ impl SuiCommand {
                 // if Sui config dir is not empty then either clean it
                 // up (if --force/-f option was specified or report an
                 // error
-                if sui_config_dir
-                    .read_dir()
-                    .map_err(|err| {
-                        anyhow!(err)
-                            .context(format!("Cannot open Sui config dir {:?}", sui_config_dir))
-                    })?
-                    .next()
-                    .is_some()
+                if write_config.is_none()
+                    && sui_config_dir
+                        .read_dir()
+                        .map_err(|err| {
+                            anyhow!(err)
+                                .context(format!("Cannot open Sui config dir {:?}", sui_config_dir))
+                        })?
+                        .next()
+                        .is_some()
                 {
                     if *force {
                         fs::remove_dir_all(sui_config_dir).map_err(|err| {
@@ -141,8 +177,13 @@ impl SuiCommand {
                     None => GenesisConfig::default_genesis(sui_config_dir)?,
                 };
 
-                let (network_config, accounts, mut keystore) = genesis(genesis_conf).await?;
+                if let Some(path) = write_config {
+                    let persisted = genesis_conf.persisted(path);
+                    persisted.save()?;
+                    return Ok(());
+                }
 
+                let (network_config, accounts, mut keystore) = genesis(genesis_conf).await?;
                 info!("Network genesis completed.");
                 let network_config = network_config.persisted(&network_path);
                 network_config.save()?;
